Fixes for 1.21

+++ common/src/main/java/com/evandev/reliable_backpacks/common/blocks/BackpackBlock.java
@@ -7,11 +7,12 @@ import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.core.Holder;
 import net.minecraft.sounds.SoundEvent;
-import net.minecraft.world.Containers;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.EquipmentSlot;
+import net.minecraft.world.entity.item.ItemEntity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.Equipable;
+import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.context.BlockPlaceContext;
 import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.Level;
@@ -27,30 +28,24 @@ import net.minecraft.world.level.block.state.properties.BooleanProperty;
 import net.minecraft.world.level.block.state.properties.DirectionProperty;
 import net.minecraft.world.level.material.FluidState;
 import net.minecraft.world.level.material.Fluids;
+import net.minecraft.world.level.storage.loot.LootParams;
+import net.minecraft.world.level.storage.loot.parameters.LootContextParams;
 import net.minecraft.world.phys.BlockHitResult;
 import net.minecraft.world.phys.shapes.CollisionContext;
 import net.minecraft.world.phys.shapes.VoxelShape;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-public class BackpackBlock extends BaseEntityBlock implements Equipable, EntityBlock, SimpleWaterloggedBlock {
-    public static final DirectionProperty FACING;
-    public static final BooleanProperty FLOATING;
-    public static final BooleanProperty WATERLOGGED;
-    protected static final VoxelShape SHAPE_X;
-    protected static final VoxelShape SHAPE_Z;
-    protected static final VoxelShape FLOATING_SHAPE_X;
-    protected static final VoxelShape FLOATING_SHAPE_Z;
+import java.util.List;
 
-    static {
-        FACING = HorizontalDirectionalBlock.FACING;
-        FLOATING = BooleanProperty.create("floating");
-        WATERLOGGED = BlockStateProperties.WATERLOGGED;
-        SHAPE_X = Block.box(3.0, 0.0, 4.0, 13.0, 11.0, 12.0);
-        SHAPE_Z = Block.box(4.0, 0.0, 3.0, 12.0, 11.0, 13.0);
-        FLOATING_SHAPE_X = Block.box(3.0, 0.0, 4.0, 13.0, 8.0, 12.0);
-        FLOATING_SHAPE_Z = Block.box(4.0, 0.0, 3.0, 12.0, 8.0, 13.0);
-    }
+public class BackpackBlock extends BaseEntityBlock implements Equipable, EntityBlock, SimpleWaterloggedBlock {
+    public static final DirectionProperty FACING = HorizontalDirectionalBlock.FACING;
+    public static final BooleanProperty FLOATING = BooleanProperty.create("floating");
+    public static final BooleanProperty WATERLOGGED = BlockStateProperties.WATERLOGGED;
+    protected static final VoxelShape SHAPE_X = Block.box(3.0, 0.0, 4.0, 13.0, 11.0, 12.0);
+    protected static final VoxelShape SHAPE_Z = Block.box(4.0, 0.0, 3.0, 12.0, 11.0, 13.0);
+    protected static final VoxelShape FLOATING_SHAPE_X = Block.box(3.0, 0.0, 4.0, 13.0, 8.0, 12.0);
+    protected static final VoxelShape FLOATING_SHAPE_Z = Block.box(4.0, 0.0, 3.0, 12.0, 8.0, 13.0);
 
     public BackpackBlock(Properties properties) {
         super(properties);
@@ -63,7 +58,6 @@ public class BackpackBlock extends BaseEntityBlock implements Equipable, EntityB
     public BlockState getStateForPlacement(BlockPlaceContext context) {
         BlockPos pos = context.getClickedPos();
         Level level = context.getLevel();
-
         return this.defaultBlockState()
                 .setValue(FACING, context.getHorizontalDirection())
                 .setValue(FLOATING, level.getFluidState(pos.below()).isSource() && !level.getFluidState(pos).isSource())
@@ -102,10 +96,8 @@ public class BackpackBlock extends BaseEntityBlock implements Equipable, EntityB
         } else {
             BlockEntity blockEntity = level.getBlockEntity(pos);
             if (blockEntity instanceof BackpackBlockEntity backpackBlockEntity) {
-
                 player.openMenu(backpackBlockEntity);
                 backpackBlockEntity.onOpen(player);
-
                 return InteractionResult.CONSUME;
             } else {
                 return InteractionResult.PASS;
@@ -114,8 +106,40 @@ public class BackpackBlock extends BaseEntityBlock implements Equipable, EntityB
     }
 
     protected void onRemove(@NotNull BlockState state, @NotNull Level level, @NotNull BlockPos pos, @NotNull BlockState newState, boolean isMoving) {
-        Containers.dropContentsOnDestroy(state, newState, level, pos);
-        super.onRemove(state, level, pos, newState, isMoving);
+        if (!state.is(newState.getBlock())) {
+            level.updateNeighbourForOutputSignal(pos, this);
+            super.onRemove(state, level, pos, newState, isMoving);
+        }
+    }
+
+    @Override
+    public @NotNull BlockState playerWillDestroy(Level level, @NotNull BlockPos pos, @NotNull BlockState state, @NotNull Player player) {
+        BlockEntity blockEntity = level.getBlockEntity(pos);
+        if (blockEntity instanceof BackpackBlockEntity backpack) {
+            if (!level.isClientSide && player.isCreative() && !backpack.isEmpty()) {
+                ItemStack itemStack = new ItemStack(this);
+                itemStack.applyComponents(backpack.collectComponents());
+                ItemEntity itemEntity = new ItemEntity(level, pos.getX() + 0.5D, pos.getY() + 0.5D, pos.getZ() + 0.5D, itemStack);
+                itemEntity.setDefaultPickUpDelay();
+                level.addFreshEntity(itemEntity);
+            }
+        }
+        return super.playerWillDestroy(level, pos, state, player);
+    }
+
+    @Override
+    public @NotNull List<ItemStack> getDrops(@NotNull BlockState state, @NotNull LootParams.Builder builder) {
+        List<ItemStack> drops = super.getDrops(state, builder);
+        BlockEntity blockEntity = builder.getOptionalParameter(LootContextParams.BLOCK_ENTITY);
+
+        if (blockEntity instanceof BackpackBlockEntity backpack) {
+            for (ItemStack drop : drops) {
+                if (drop.is(this.asItem())) {
+                    drop.applyComponents(backpack.collectComponents());
+                }
+            }
+        }
+        return drops;
     }
 
     protected @NotNull VoxelShape getShape(BlockState state, @NotNull BlockGetter level, @NotNull BlockPos pos, @NotNull CollisionContext context) {
@@ -125,7 +149,6 @@ public class BackpackBlock extends BaseEntityBlock implements Equipable, EntityB
         } else {
             return direction.getAxis() == Direction.Axis.X ? SHAPE_Z : SHAPE_X;
         }
-
     }
 
     @Override
